backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
pillow==10.3.0
imagehash==4.3.1
sentence-transformers==3.0.1
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
python-multipart==0.0.9
aiofiles==23.2.1
